Add files via upload
The "PhotoGalleries" project is a web application designed to showcase photo galleries with a responsive and visually appealing interface. It utilizes HTML, CSS (Bootstrap), and JavaScript, incorporating libraries like AOS for animations, GLightbox for lightbox functionality, and Swiper for image sliders. The project includes a structured layout with documentation, source code, vendor assets, and various images for galleries, profiles, and testimonials. Additionally, it features a feedback button that allows users to submit feedback, which is connected to a backend SQL Server for storing and managing the responses. Ideal for photography portfolios, event galleries, and product showcases, this application provides a seamless and engaging user experience across different devices and screen sizes